add bgs / fgs / merger, extend augmentator

+++ code/utils/augment.py
@@ -23,7 +23,8 @@ class YoloAugmentator:
         files_to_ignore: List[str],
         rot_transition: dict,
         flip_transition: dict,
-        perform_augmentation: bool
+        perform_augmentation: bool,
+        clean: bool = True
     ):
         self.label_dir = label_dir
         self.preprocessed_dir = preprocessed_dir
@@ -32,6 +33,7 @@ class YoloAugmentator:
         self.rot_transition = rot_transition
         self.flip_transition = flip_transition
         self.perform_augmentation = perform_augmentation
+        self.clean = clean
 
     def augment(self, file: str, oimg, ocontent):
         # rotate original image
@@ -245,7 +247,8 @@ class YoloAugmentator:
             try:
                 imgs_to_augment.pop(imgs_to_augment.index(file))
             except Exception as e:
-                print(e)
+                # print(e)
+                pass
 
         # exclude already augmented files
         # files_to_augment = [f for f in files_to_augment if not "aug" in f]
@@ -255,8 +258,9 @@ class YoloAugmentator:
         return imgs_to_augment
 
     def run(self):
-        for filename in os.listdir(self.preprocessed_dir):
-            os.remove(self.preprocessed_dir / filename)
+        if self.clean:
+            for filename in os.listdir(self.preprocessed_dir):
+                os.remove(self.preprocessed_dir / filename)
 
         # os.symlink(
         #     os.path.abspath(self.label_dir / "classes.txt"),
@@ -505,11 +509,24 @@ if __name__ == "__main__":
             files_to_ignore,
             label_transition_rotation,
             label_transition_flip,
-            config.augment.perform_train
+            config.augment.perform_train,
+            clean=True
         )
         augmentator.run()
 
-        print("Stripping classes from train_preprocessed...")
+        print("Augmenting train_merged files, this may take some time...")
+        augmentator = YoloAugmentator(
+            config.merged_dir,
+            config.preprocessed_dir,
+            files_to_ignore,
+            label_transition_rotation,
+            label_transition_flip,
+            config.augment.perform_merged,
+            clean=False
+        )
+        augmentator.run()
+
+        print("Stripping classes from train preprocessed...\n")
         ClassStripper(
             config.preprocessed_dir,
             config.labels_to_remove,
@@ -528,7 +545,8 @@ if __name__ == "__main__":
             files_to_ignore,
             label_transition_rotation,
             label_transition_flip,
-            config.augment.perform_valid
+            config.augment.perform_valid,
+            clean=True
         )
         augmentator.run()
